perf: Activity.fixSomeProblems
