+++ app/src/main/kotlin/li/songe/gkd/util/Others.kt
@@ -1,7 +1,14 @@
 package li.songe.gkd.util
 
+import android.annotation.SuppressLint
+import android.app.Activity
 import android.content.ComponentName
+import android.content.res.Resources
 import android.graphics.Bitmap
+import android.graphics.Color
+import android.os.Build
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
 import li.songe.gkd.META
 import kotlin.reflect.KClass
 import kotlin.reflect.jvm.jvmName
@@ -38,5 +45,45 @@ fun getShowActivityId(appId: String, activityId: String?): String? {
         }
     } else {
         null
+    }
+}
+
+fun Activity.fixSomeProblems() {
+    fixTopPadding()
+    fixTransparentNavigationBar()
+}
+
+private fun Activity.fixTopPadding() {
+    // 当调用系统分享时, 会导致状态栏区域消失, 应用整体上移, 设置一个 top padding 保证不上移
+    // 当系统存在工作空间时, 分享界面展示不一样, 不会产生上移行为
+    var tempTop = Resources.getSystem().run {
+        @SuppressLint("DiscouragedApi", "InternalInsetResource")
+        getDimensionPixelSize(getIdentifier("status_bar_height", "dimen", "android"))
+    }
+    ViewCompat.setOnApplyWindowInsetsListener(window.decorView) { view, windowInsets ->
+        view.setBackgroundColor(Color.TRANSPARENT)
+        val statusBars = windowInsets.getInsets(WindowInsetsCompat.Type.statusBars())
+        if (statusBars.top == 0) {
+            view.setPadding(
+                statusBars.left,
+                tempTop,
+                statusBars.right,
+                statusBars.bottom
+            )
+        } else {
+            tempTop = statusBars.top
+            view.setPadding(statusBars.left, 0, statusBars.right, statusBars.bottom)
+        }
+        ViewCompat.onApplyWindowInsets(view, windowInsets)
+    }
+}
+
+private fun Activity.fixTransparentNavigationBar() {
+    // 修复在浅色主题下导航栏背景不透明的问题
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+        window.isNavigationBarContrastEnforced = false
+    } else {
+        @Suppress("DEPRECATION")
+        window.navigationBarColor = Color.TRANSPARENT
     }
 }
